// types/cdp.ts
export const REQUEST = "request" as const;
export const RESPONSE = "response" as const;
export const EVENT = "event" as const;

export { zod, commands, events };
export const cdp = { types: runtimeTypes, commands, events, REQUEST, RESPONSE, EVENT } as const;
export const types = cdp.types;

export namespace cdp {
  export namespace types {
    export namespace ts {
      export namespace Mod {
        export type Routes = CDPModRoutes;
        export type CustomPayload = CDPModCustomPayload;
        export type NamedValue = CDPModNamedValue;
        export type Name = CDPModName;
        export type ZodType = CDPModZodType;
        export type PayloadShape = CDPModPayloadShape;
        export type PayloadSchemaSpec = CDPModPayloadSchemaSpec;
        export type EvaluateParams = CDPModEvaluateParams;
// ...
      }
// ...
      export namespace Runtime {
// ...
        export type EvaluateResult = z.infer<typeof zod.Runtime.EvaluateResult>;
// ...
      }
// ...
    }
  }
}